+++ cypress/tests/open-ebooks/integration/login.ts
@@ -2,7 +2,8 @@ import {
   APP_PATH,
   COOKIE_KEY,
   FIRSTBOOK_LOGIN_PATH,
-  LOGIN_PATH
+  LOGIN_PATH,
+  SERVER_URL
 } from "../../../support/utils";
 
 describe("Open eBooks login flow", () => {
@@ -52,6 +53,10 @@ describe("FirstBook login", () => {
   });
 
   it("Can login with FirstBook", () => {
+    cy.intercept("GET", `${SERVER_URL}/groups`, {
+      fixture: "open-ebooks/all-access/all-collections.html"
+    }).as("allCollections");
+
     cy.visit("/");
     cy.findByRole("link", { name: "Login with First Book" }).click();
     cy.location("pathname").should("eq", FIRSTBOOK_LOGIN_PATH);
@@ -68,7 +73,15 @@ describe("FirstBook login", () => {
     cy.findByRole("button", { name: "Login" }).click();
 
     cy.log("wait for url to change after submitting credentials");
-    cy.url().should("eq", `${Cypress.config().baseUrl}${APP_PATH}`);
+    cy.url({ timeout: 7000 }).should(
+      "eq",
+      `${Cypress.config().baseUrl}${APP_PATH}`
+    );
+    cy.wait("@allCollections");
+
+    cy.log(
+      "verify you are on the landing page and collection lanes are present"
+    );
     cy.get("main")
       .findByRole("heading", { name: "Open eBooks (QA Server) Home" })
       .should("exist");
@@ -80,19 +93,6 @@ describe("FirstBook login", () => {
   it("Redirects to collection page after login", () => {
     // FIXME: Redirect is not happening when first navigating to a collection and then logging in
     // cy.visit(ALL_ACCESS_AUTHOR_RECOMMENDATIONS_PATH_HUNTER_C_C);
-    // cy.findByRole("link", { name: "Sign In" }).click();
-    // cy.findByRole("link", { name: "Login with First Book" }).click();
-    // cy.log("type username and password");
-    // cy.findByRole("textbox", { name: "Access Code input" }).type(
-    //   Cypress.env("OPENEBOOKS_ALL_ACCESS_USER_USERNAME"),
-    //   { log: false }
-    // );
-    // cy.findByLabelText("PIN input").type(
-    //   Cypress.env("OPENEBOOKS_ALL_ACCESS_USER_PW"),
-    //   { log: false }
-    // );
-    // cy.findByRole("button", { name: "Login" }).click();
-    // cy.findByRole("heading", { name: "Hunter, C. C." }).should("exist");
   });
 
   it("Displays error when entering nothing", () => {
@@ -128,26 +128,19 @@ describe("FirstBook login", () => {
       Cypress.env("OPENEBOOKS_ALL_ACCESS_USER_USERNAME"),
       { log: false }
     );
-    cy.findByLabelText("PIN input").type("1234", { log: false });
+    cy.findByLabelText("PIN input").type("0987234", { log: false });
+
+    cy.intercept("GET", `${SERVER_URL}/loans`, {
+      statusCode: 401,
+      fixture: "open-ebooks/incorrect-credentials.json"
+    }).as("incorrectCredentials");
+
     cy.findByRole("button", { name: "Login" }).click();
+    cy.wait("@incorrectCredentials")
+      .its("response.statusCode")
+      .should("eq", 401);
     cy.findByText(
       "Invalid credentials: A valid library card barcode number and PIN are required."
     ).should("exist");
-  });
-
-  it("Can logout with FirstBook", () => {
-    cy.loginByApi("ALL_ACCESS_USER");
-    cy.visit(APP_PATH);
-    cy.wait(5000);
-    cy.findByRole("button", { name: "Sign Out" }).should("exist").click();
-    cy.findByRole("button", { name: "Cancel" }).should("exist");
-    cy.findByRole("button", { name: "Confirm Sign Out" })
-      .should("exist")
-      .click();
-    cy.wait(5000);
-    cy.location("pathname").should("eq", LOGIN_PATH);
-    cy.findByRole("listitem", { name: "Current location: Login" }).should(
-      "exist"
-    );
   });
 });